LLMOps/LLM_Tracking_mlflow/pyproject.toml
[project]
name = "llm-tracking-mlflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-ibm>=1.0.4",
    "mlflow>=3.10.1",
    "python-dotenv>=1.2.2",
]
